refactoring: child-components moved to separate folders inside SubTaskList
