+++ src/components/SubTaskList/SubTaskItem/SubTaskItem.jsx
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import {
   Button, Form, Row, Col,
 } from 'react-bootstrap';
+import './SubTaskItem.scss';
 
 export default class Subtask extends Component {
   render() {
